// gen/ts/open_splunk/collector_admin.ts
export enum IngestionTokenState {
  INGESTION_TOKEN_STATE_UNSPECIFIED = 0,
  INGESTION_TOKEN_STATE_ACTIVE = 1,
  INGESTION_TOKEN_STATE_DISABLED = 2,
  INGESTION_TOKEN_STATE_REVOKED = 3,
  INGESTION_TOKEN_STATE_EXPIRED = 4,
  UNRECOGNIZED = -1,
}
// ...
export enum IngestionTokenPurpose {
  INGESTION_TOKEN_PURPOSE_UNSPECIFIED = 0,
  INGESTION_TOKEN_PURPOSE_NATIVE_COLLECTOR = 1,
  INGESTION_TOKEN_PURPOSE_HEC = 2,
  UNRECOGNIZED = -1,
}